Move all Footer subcomponents inside a folder
