+++ packages/react/src/components/footer/Footer.tsx
@@ -6,14 +6,14 @@ import '../../styles/base.css';
 import styles from './Footer.module.scss';
 import { Koros, KorosType } from '../koros';
 import classNames from '../../utils/classNames';
-import { FooterNavigation } from './footerNavigation/FooterNavigation';
-import { FooterNavigationGroup } from './footerNavigationGroup/FooterNavigationGroup';
-import { FooterGroupHeading } from './footerGroupHeading/FooterGroupHeading';
-import { FooterNavigationLink } from './footerNavigationLink/FooterNavigationLink';
-import { FooterUtilities } from './footerUtilities/FooterUtilities';
-import { FooterUtilityGroup } from './footerUtilityGroup/FooterUtilityGroup';
-import { FooterBase } from './footerBase/FooterBase';
-import { FooterCustom } from './footerCustom/FooterCustom';
+import { FooterNavigation } from './components/footerNavigation/FooterNavigation';
+import { FooterNavigationGroup } from './components/footerNavigationGroup/FooterNavigationGroup';
+import { FooterGroupHeading } from './components/footerGroupHeading/FooterGroupHeading';
+import { FooterNavigationLink } from './components/footerNavigationLink/FooterNavigationLink';
+import { FooterUtilities } from './components/footerUtilities/FooterUtilities';
+import { FooterUtilityGroup } from './components/footerUtilityGroup/FooterUtilityGroup';
+import { FooterBase } from './components/footerBase/FooterBase';
+import { FooterCustom } from './components/footerCustom/FooterCustom';
 import { FooterTheme } from './Footer.interface';
 import { useTheme } from '../../hooks/useTheme';
