+++ packages/react/src/components/footer/components/footerBase/FooterBase.module.scss
@@ -1,0 +1,105 @@
+@import '../../common.scss';
+@value x-small-down, small-down, small-up, medium-down, medium-up, large-up from "../../../../styles/breakpoints.scss";
+
+.backToTopButton {
+  @extend %footerItem;
+
+  align-self: unset;
+  font-size: 20px;
+  font-weight: bold;
+  line-height: 30px;
+  margin-left: auto;
+}
+
+.base {
+  display: flex;
+  flex-direction: column;
+  padding-bottom: var(--spacing-m);
+}
+
+.baseActions {
+  display: flex;
+  flex-grow: 1;
+}
+
+.copyright {
+  display: flex;
+  font-size: var(--fontsize-body-s);
+  gap: var(--spacing-3-xs);
+  justify-content: flex-start;
+  line-height: var(--lineheight-l);
+  margin-right: var(--spacing-2-xs);
+}
+
+.divider {
+  @extend %divider;
+}
+
+.links {
+  align-items: center;
+  display: flex;
+  gap: var(--spacing-xs);
+}
+
+.logoWrapper {
+  display: flex;
+}
+
+.separator {
+  font-size: var(--fontsize-body-s);
+  font-weight: 500;
+  line-height: var(--lineheight-l);
+}
+
+@media x-small-down {
+  .links {
+    align-items: flex-start;
+    flex-direction: column;
+  }
+  
+  .copyright {
+    &Dot {
+      display: none;
+    }
+  }
+  
+  .separator {
+    display: none;
+  }
+}
+
+@media small-down {
+  .baseActions {
+    flex-direction: column;
+  }
+
+  .backToTopButton {
+    margin-top: var(--spacing-m);
+  }
+}
+
+@media medium-down {
+  .copyright {
+    margin-bottom: var(--spacing-xs);
+    margin-top: var(--spacing-s);
+  }
+
+  .divider {
+    margin: 0 0 var(--spacing-m);
+  }
+  
+  .separator:first-of-type {
+    display: none;
+  }
+}
+
+@media large-up {
+  .base {
+    align-items: center;
+    flex-flow: row wrap;
+  }
+
+  .copyright {
+    margin-left: var(--spacing-s);
+  }
+}
