+++ packages/react/src/components/footer/components/footerCustom/FooterCustom.module.scss
@@ -1,0 +1,11 @@
+@import '../../common.scss';
+
+.custom {
+  padding-bottom: var(--spacing-m);
+}
+
+.divider {
+  @extend %divider;
+  
+  margin: 0 0 var(--spacing-m);
+}
