+++ packages/react/src/components/footer/components/footerUtilityGroup/FooterUtilityGroup.module.scss
@@ -1,0 +1,47 @@
+@value x-small-down, small-only, medium-up, large-up, x-large-up from "../../../../styles/breakpoints.scss";
+
+.utilityGroup {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-xs);
+  justify-items: start;
+  list-style: none;
+  margin-bottom: var(--spacing-xs);
+  padding: 0;
+  width: 248px;
+}
+
+@media x-small-down {
+  .utilityGroup:nth-last-child(-n+1) {
+    margin-bottom: 0;
+  }
+}
+
+@media small-only {
+  .utilityGroup:nth-last-child(-n+2) {
+    margin-bottom: 0;
+  }
+}
+
+@media medium-up {
+  .utilityGroup {
+    width: 214px
+  }
+
+  .utilityGroup:nth-last-child(-n+1) {
+    margin-bottom: 0;
+  }
+}
+
+@media large-up {
+  .utilityGroup {
+    margin-bottom: 0;
+    width: 204px;
+  }
+}
+
+@media x-large-up {
+  .utilityGroup {
+    width: 268px
+  }
+}
