--- packages/react/src/components/footer/footerCustom/FooterCustom.module.scss
@@ -1,11 +1,0 @@
-@import '../common.scss';
-
-.custom {
-  padding-bottom: var(--spacing-m);
-}
-
-.divider {
-  @extend %divider;
-  
-  margin: 0 0 var(--spacing-m);
-}
