--- packages/react/src/components/footer/footerNavigationGroup/FooterNavigationGroup.module.scss
@@ -1,22 +1,0 @@
-@value x-large-up from "../../../styles/breakpoints.scss";
-
-.navigationGroup {
-  display: flex;
-}
-
-.navigationGroupList {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacing-3-xs);
-  justify-items: start;
-  list-style: none;
-  margin: 0;
-  padding: 0;
-  width: 137px;
-}
-
-@media x-large-up {
-  .navigationGroupList {
-    width: 180px;
-  }
-}
