--- packages/react/src/components/footer/footerNavigationLink/FooterNavigationLink.module.scss
@@ -1,61 +1,0 @@
-@import "../common.scss";
-@value medium-down, large-down from "../../../styles/breakpoints.scss";
-
-.item {
-  @extend %footerItem;
-  
-  font-weight: normal;
-  line-height: var(--lineheight-l);
-}
-
-.base {
-  align-self: unset;
-  font-size: var(--fontsize-body-s);
-  font-weight: 500;
-  margin: 0;
-  text-decoration: underline;
-}
-
-.icon {
-  composes: icon from 'hds-core/lib/components/link/link.css';
-  margin-left: var(--spacing-3-xs);
-}
-
-.verticalAlignSmallOrMediumIcon {
-  composes: vertical-align-small-or-medium-icon from 'hds-core/lib/components/link/link.css';
-}
-
-.navigation {
-  align-items: center;
-  font-size: var(--fontsize-body-l);
-}
-
-.utility {
-  align-items: center;
-  font-size: var(--fontsize-body-l);
-}
-
-.subItem.subItem {
-  font-size: var(--fontsize-body-s);
-  line-height: var(--lineheight-l)
-}
-
-.subItemIcon {
-  align-self: start;
-  height: var(--spacing-xs);
-  min-width: var(--spacing-xs);
-  transform: translateY(4px);
-  width: var(--spacing-xs);
-}
-
-@media medium-down {
-  .utility, .navigation {
-    font-size: var(--fontsize-body-m)
-  }
-}
-
-@media large-down {
-  .item:not(.base):not(.navigation):not(.utility) {
-    font-size: var(--fontsize-body-m);
-  }
-}
