--- packages/react/src/components/footer/footerUtilities/FooterUtilities.module.scss
@@ -1,86 +1,0 @@
-@import '../common.scss';
-@value x-small-down, small-down, medium-down, large-up from "../../../styles/breakpoints.scss";
-
-.divider {
-  @extend %divider;
-}
-
-.groups {
-  display: flex;
-  flex-wrap: wrap;
-  row-gap: var(--spacing-m);
-}
-
-.links {
-  display: flex;
-  flex-wrap: wrap;
-  gap: var(--spacing-xs) var(--spacing-m);
-  max-width: 60%;
-}
-
-.soMe {
-  display: grid;
-  gap: var(--spacing-xs);
-  grid-auto-flow: column;
-
-  > * {
-    align-self: center;
-  }
-}
-
-.utilities {
-  align-items: center;
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: space-between;
-  padding-bottom: var(--spacing-m);
-}
-
-.utilityLinks {
-  display: grid;
-  grid-auto-flow: column;
-  position: relative;
-}
-
-.widerLinks {
-  max-width: none;
-}
-
-@media x-small-down {
-  .links {
-    flex-direction: column;
-    gap: var(--spacing-xs);
-    justify-items: flex-start;
-  }
-
-  .utilities {
-    align-items: flex-start;
-    flex-direction: column;
-  }
-}
-
-@media small-down {
-  .divider {
-    margin: 0 0 var(--spacing-xs);
-  }
-
-  .utilities {
-    justify-items: flex-start;
-  }
-}
-
-@media medium-down {
-  .link {
-    gap: var(--spacing-xs);
-  }
-
-  .utilities {
-    gap: var(--spacing-xs);
-  }
-}
-
-@media large-up {
-  .divider {
-    margin: 0 0 var(--spacing-m);
-  }
-}
